+++ metadata.rb
@@ -5,3 +5,8 @@ license          'All rights reserved'
 description      'Installs/Configures nuaavee development environment'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
 version          '0.1.0'
+license          'MIT'
+
+recipe 'nuaavee' 'Installs nuaavee environment configuration files'
+
+supports 'ubuntu', '= 14.04'