+++ CHANGELOG.md
@@ -5,9 +5,7 @@ This file is used to list changes made in each version of the nuaavee cookbook.
 
 0.1.0
 -----
-- [your_name] - Initial release of nuaavee
+- [anshul] - Initial release of nuaavee
 
 - - -
-Check the [Markdown Syntax Guide](http://daringfireball.net/projects/markdown/syntax) for help with Markdown.
 
-The [Github Flavored Markdown page](http://github.github.com/github-flavored-markdown/) describes the differences between markdown on github and standard markdown.